Refactor SessionStatus
