+++ Auth/src/commonMain/kotlin/io/github/jan/supabase/auth/mfa/MfaApi.kt
@@ -2,8 +2,8 @@ package io.github.jan.supabase.auth.mfa
 
 import io.github.jan.supabase.auth.Auth
 import io.github.jan.supabase.auth.AuthImpl
-import io.github.jan.supabase.auth.SessionStatus
 import io.github.jan.supabase.auth.providers.builtin.Phone
+import io.github.jan.supabase.auth.status.SessionStatus
 import io.github.jan.supabase.auth.user.UserMfaFactor
 import io.github.jan.supabase.auth.user.UserSession
 import io.github.jan.supabase.putJsonObject
