+++ Auth/src/commonTest/kotlin/AuthTestUtils.kt
@@ -1,5 +1,5 @@
 import io.github.jan.supabase.auth.Auth
-import io.github.jan.supabase.auth.SessionStatus
+import io.github.jan.supabase.auth.status.SessionStatus
 
 fun Auth.sessionSource() = (sessionStatus.value as SessionStatus.Authenticated).source
