+++ buildSrc/src/main/kotlin/KotlinTargets.kt
@@ -1,8 +1,8 @@
 import org.gradle.kotlin.dsl.assign
 import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
+import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
 import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-import org.jetbrains.kotlin.gradle.targets.js.dsl.ExperimentalWasmDsl
 
 fun KotlinMultiplatformExtension.iosTargets() {
     iosX64()
@@ -67,6 +67,7 @@ fun KotlinMultiplatformExtension.jsTarget() {
     }
 }
 
+@OptIn(org.jetbrains.kotlin.gradle.ExperimentalWasmDsl::class)
 fun KotlinMultiplatformExtension.wasmJsTarget() {
     wasmJs {
         browser {
