+++ sample/chat-demo-mpp/common/src/commonMain/kotlin/io/github/jan/supabase/common/App.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import io.github.jan.supabase.auth.SessionStatus
+import io.github.jan.supabase.auth.status.SessionStatus
 import io.github.jan.supabase.common.ui.screen.ChatScreen
 import io.github.jan.supabase.common.ui.screen.LoginScreen
